+++ src/org/bccvl/compute/brt.py
@@ -6,6 +6,8 @@ from org.bccvl.compute.utils import (prepare_data, init_work_env, addFile,
 import shutil
 from pkg_resources import resource_string
 import glob
+from plone.app.uuid.utils import uuidToObject
+
 
 BRT_CONFIG = """
 .libPaths("{rlibdir}")
@@ -64,7 +66,7 @@ def write_brt_config(rootpath, path, species):
 
     currentfolder = get_datapath_for_glob(path, 'current*')
     futurefolder = get_datapath_for_glob(path, '*2085')
-    curdata = ",".join(('"{0}"'.format(os.path.join(currentfolder, name + ".tif")) for name in names)),
+    curdata = ",".join(('"{0}"'.format(os.path.join(currentfolder, name + ".tif")) for name in names))
     futdata = ''
     if futurefolder:
         futdata = ",".join(('"{0}"'.format(os.path.join(futurefolder, name + ".tif")) for name in names))
@@ -113,27 +115,33 @@ def execute(experiment):
     names = ["bioclim_01", "bioclim_04", "bioclim_05",
              "bioclim_06", "bioclim_12", "bioclim_15",
              "bioclim_16", "bioclim_17"]
-    species = experiment.species_occurrence_dataset
+    species = uuidToObject(experiment.species_occurrence_dataset)
     #absence = experiment.species_absence_dataset
-    climate = experiment.climate_dataset
-    path = init_work_env(rootpath, species.id)
-
-    prepare_data(path, names, climate, species)
-    script = write_brt_config(rootpath, path, species.id)
-    scriptout = script + "out"
-    cmd = ['R', 'CMD', 'BATCH', '--no-save', '--no-restore', script, scriptout]
-    ret = call(cmd, shell=False)
-    # TODO: check ret for error
-    # TODO: make sure script returns proper error codes
-    # TODO: zip result and store on experiment
-    with zipfile.ZipFile(os.path.join(path, 'output.zip'), 'w', zipfile.ZIP_DEFLATED) as zipf:
-        for fname in os.listdir(os.path.join(path, 'output_brt')):
-            zipf.write(os.path.join(path, 'output_brt', fname), fname)
-        zipf.write(os.path.join(path, 'brt.Rout'), 'brt.Rout')
-    addFile(experiment,
-            filename=u'file://' + os.path.join(path, 'output.zip'),
-            mimetype='application/zip')
-    shutil.rmtree(path)
+    climate = uuidToObject(experiment.environmental_dataset)
+    try:
+        path = init_work_env(rootpath, species.id)
+        prepare_data(path, names, climate, species)
+        script = write_brt_config(rootpath, path, species.id)
+        # TODO: use script and scriptout instead of hardcoded brt.Rout etc...
+        scriptout = script + "out"
+        cmd = ['R', 'CMD', 'BATCH', '--no-save', '--no-restore', script, scriptout]
+        ret = call(cmd, shell=False)
+        # TODO: check ret for error
+        # TODO: make sure script returns proper error codes
+        # TODO: zip result and store on experiment
+        with zipfile.ZipFile(os.path.join(path, 'output.zip'), 'w', zipfile.ZIP_DEFLATED) as zipf:
+            for fname in os.listdir(os.path.join(path, 'output_brt')):
+                zipf.write(os.path.join(path, 'output_brt', fname), fname)
+            zipf.write(os.path.join(path, 'brt.Rout'), 'brt.Rout')
+        addFile(experiment,
+                # TODO: need IStorage adapter for urllib.urlinfo
+                #filename=u'file://' + os.path.join(path, 'output.zip'),
+                filename=os.path.join(path, 'output.zip'),
+                mimetype='application/zip')
+    finally:
+        # TODO: capture detailed error message to report to user
+        if os.path.exists(path):
+            shutil.rmtree(path)
 
 
 if __name__ == '__main__':
